move album array and interface into shared folder and file to be used globally
